refactor (Racing.java): 서비스 클래스는 무상태여야 하므로 도메인으로 이동
