+++ src/main/java/racinggame/domain/racing/Racing.java
@@ -1,7 +1,6 @@
-package racinggame.servcie;
+package racinggame.domain.racing;
 
 import racinggame.domain.car.Cars;
-import racinggame.domain.racing.RacingRoundResults;
 import racinggame.util.NumberUtil;
 
 import java.util.ArrayList;
